Adds model graph, refreshes view from it
Adds Requestant (a person and hiso'er requested days of various levels).
Desires holds a list of Requestants and a map of Days to a List
of Requestants for that day. Refreshes the view from said map. Enables
adding a person and desire semi-sensibly. Semi because this still
doesn't constrain, but it checks for a conflict in seniority.

Updated readme photo with the 'swap person with same seniority' dialog