Simple cases for constrainer
Constrainer handles the simple cases for adding or removing requests.
(Add: spot open or is an inactive level ; remove: spot open). Otherwise,
constrainer will put them in from scratch. I would prefer to start with
heuristics, but this is an mvp. Unless it is terrible (which it may be),
I'm satisfied that this is correct. I can improve it afterward.

Requestants have seniority and Desires updates that when creating a
new person or moving himo'er. The alternative is for Constrainer to
receive a reference to Desires so it can report the seniority.
It's an even toss.